+++ include/unicode.h
@@ -11,8 +11,8 @@ extern "C" {
 #endif
 
 #include <stdbool.h>
+#include <stddef.h>
 #include <stdint.h>
-#include <stdio.h>
 
 #define UNICODE_VERSION 0x0100          // 1.0
 #define UNICODE_MAX_CODEPOINT 0x10FFFF  // 1,114,111
@@ -51,7 +51,8 @@ void utf8_remove(utf8_string* s, size_t index, size_t count);
 void utf8_replace(utf8_string* s, const char* old_str, const char* new_str);
 void utf8_replace_all(utf8_string* s, const char* old_str, const char* new_str);
 void utf8_reverse(utf8_string* s);
-ssize_t utf8_writeto(const utf8_string* s, const char* filename);
+
+unsigned long utf8_writeto(const utf8_string* s, const char* filename);
 utf8_string* utf8_readfrom(const char* filename);
 
 // Split a string into parts using a delimiter.
